modified some group related stuff.  removed comments from arrays instrumentation classes.

+++ src/java/org/deuce/distribution/replication/partitioner/group/RandomGroupPartitioner.java
@@ -52,7 +52,7 @@ public class RandomGroupPartitioner extends Partitioner implements
 		List<Group> g = super.getGroups();
 		for (int i = 0; i < groups; i++)
 		{
-			g.add(new PartialReplicationGroup(i)); // XXX
+			g.add(new PartialReplicationGroup((short) i)); // XXX
 		}
 
 		for (Address a : members)
