+++ src/java/org/deuce/distribution/replication/partitioner/group/RoundRobinGroupPartitioner.java
@@ -32,7 +32,7 @@ public class RoundRobinGroupPartitioner extends Partitioner implements
 		List<Group> g = super.getGroups();
 		for (int i = 0; i < groups; i++)
 		{
-			g.add(new PartialReplicationGroup(i)); // XXX
+			g.add(new PartialReplicationGroup((short) i)); // XXX
 		}
 
 		int group = 0;
